Basics--Simplify foldl, scanl explanations
Basics.md, clearer presentation of foldl and scanl.
Working from master branch.  9 Sep 2020 4:20 p.m.

The explanation of `foldl` and `scanl` could be made simpler.

First, the easiest example should be shown first, which is the addition example (`(+)`), because everyone is really familiar with addition.  If you see the addition example, you can picture it in your head, whereas the others are not as easy to visualize.  Granted, addition does not give as much of a feel for how the one value is being transitioned, and the input sequence of values is being used to transition it, but that can be learned from the harder examples.

Second, even though it seems we are more interested in learning `foldl` than `scanl`, the better order to learn them is to start with scanl.  If you do `foldl` before `scanl`, then for `foldl` you have to imagine stepping through the sequence and getting the final value, and then for `scanl` you have to retrace your steps to imagine outputting the whole sequence.  But if you do `scanl` before `foldl`, then you get the whole sequence with `scanl`, and `foldl` requires no more work--just grab the last value.

Also, by the way, the existing explanation defines `foldl` partway, without saying what the output is at first, and then defines `scanl`, and while defining `scanl` finally says what the output of `foldl` is.  So that should be fixed at the same time.

Hence the suggested changes.